+++ packages/collector/test/tracing/cloud/aws/dynamodb/app.js
@@ -183,7 +183,7 @@ const DynamoDBApi = {
 };
 
 const app = express();
-const port = process.env.APP_SENDER_PORT || 3215;
+const port = process.env.APP_PORT || 3215;
 
 app.get('/', (_req, res) => {
   res.send('Ok');