+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Instrumentation of memcached
+
 ## 1.125.1
 - Reporting correct SQS entry span duration
 - Reporting SQS-Consumer processing error messages to the agent